Ensure output is always a valid nextflow config
Error messages are now printed to stderr, but the config generated on
stdout will also raise an exception to Nextflow that contains the same
error message. This will eliminate the cryptic Nextflow errors like
"Unknown method invocation `Raw` on _parse_closure5 type" that were
caused by writing a plain text error string to the config. Ultimately
the calling scripts should `set -e` so a failure in this script kills
the whole thing rather than proceeding with a broken config.